# Cost Problem
Earning by Multitouching](https://www.notion.so/Earning-by-Multitouching-ba9a7decbb6e4ee6848bb10fa580cffd)

[Earning Count](https://www.notion.so/Earning-Count-2a89c82118c04780a4996cba024982c2)

+++ Assets/Project/Scripts/Managers/GameManager.cs
@@ -380,24 +380,27 @@ public class GameManager : MonoSingleton<GameManager>
 
     void Update()
     {
-        if (Input.GetKeyDown(KeyCode.M)||Input.GetMouseButtonDown(1))
+        if(Application.isEditor)
         {
-            gold += 1000;
-            PlayerPrefs.SetInt("Gold", gold);
-            UIManager.Instance.goldText.text = "" + gold;
-            UIManager.Instance.UpdateEconomyUI();
-        }
+            if (Input.GetKeyDown(KeyCode.M) || Input.GetMouseButtonDown(1))
+            {
+                gold += 1000;
+                PlayerPrefs.SetInt("Gold", gold);
+                UIManager.Instance.goldText.text = "" + gold;
+                UIManager.Instance.UpdateEconomyUI();
+            }
 
-        if (Input.GetKeyDown(KeyCode.R) || Input.GetMouseButtonDown(3))
-        {
-            PlayerPrefs.DeleteAll();
-            Application.LoadLevel(0);
-        }
+            if (Input.GetKeyDown(KeyCode.R) || Input.GetMouseButtonDown(3))
+            {
+                PlayerPrefs.DeleteAll();
+                Application.LoadLevel(0);
+            }
 
-        if (Input.GetKeyDown(KeyCode.S) || Input.GetMouseButtonDown(2))
-            simulationSpeed = 10;
-        else if (Input.GetKeyUp(KeyCode.S) || Input.GetMouseButtonUp(2))
-            simulationSpeed = 1;
+            if (Input.GetKeyDown(KeyCode.S) || Input.GetMouseButtonDown(2))
+                simulationSpeed = 10;
+            else if (Input.GetKeyUp(KeyCode.S) || Input.GetMouseButtonUp(2))
+                simulationSpeed = 1;
+        }
     }
 
 #endregion
